requirements.txt
attrs==21.2.0
aws-cdk.assertions==1.134.0
aws-cdk.assets==1.134.0
aws-cdk.aws-acmpca==1.134.0
aws-cdk.aws-apigateway==1.134.0
aws-cdk.aws-applicationautoscaling==1.134.0
aws-cdk.aws-autoscaling-common==1.134.0
aws-cdk.aws-certificatemanager==1.134.0
aws-cdk.aws-cloudformation==1.134.0
aws-cdk.aws-cloudwatch==1.134.0
aws-cdk.aws-codeguruprofiler==1.134.0
aws-cdk.aws-codestarnotifications==1.134.0
aws-cdk.aws-cognito==1.134.0
aws-cdk.aws-dynamodb==1.134.0
aws-cdk.aws-ec2==1.134.0
aws-cdk.aws-ecr==1.134.0
aws-cdk.aws-ecr-assets==1.134.0
aws-cdk.aws-efs==1.134.0
aws-cdk.aws-elasticloadbalancingv2==1.134.0
aws-cdk.aws-events==1.134.0
aws-cdk.aws-iam==1.134.0
aws-cdk.aws-kinesis==1.134.0
aws-cdk.aws-kms==1.134.0
aws-cdk.aws-lambda==1.134.0
aws-cdk.aws-logs==1.134.0
aws-cdk.aws-route53==1.134.0
aws-cdk.aws-s3==1.134.0
aws-cdk.aws-s3-assets==1.134.0
aws-cdk.aws-signer==1.134.0
aws-cdk.aws-sns==1.134.0
aws-cdk.aws-sqs==1.134.0
aws-cdk.aws-ssm==1.134.0
aws-cdk.cloud-assembly-schema==1.134.0
aws-cdk.core==1.134.0
aws-cdk.custom-resources==1.134.0
aws-cdk.cx-api==1.134.0
aws-cdk.region-info==1.134.0
boto3==1.20.14
botocore==1.23.14
cattrs==1.8.0
constructs==3.3.161
jmespath==0.10.0
jsii==1.46.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
